+++ finagle-core/src/main/scala/com/twitter/finagle/client/MethodBuilder.scala
@@ -5,6 +5,8 @@ import com.twitter.finagle.client.MethodBuilderTimeout.TunableDuration
 import com.twitter.finagle.service.Filterable
 import com.twitter.finagle.service.ResponseClass
 import com.twitter.finagle.service.ResponseClassifier
+import com.twitter.finagle.service.Retries
+import com.twitter.finagle.service.RetryBudget
 import com.twitter.finagle.service.TimeoutFilter
 import com.twitter.finagle.stats.LazyStatsReceiver
 import com.twitter.finagle.stats.StatsReceiver
@@ -18,6 +20,7 @@ import com.twitter.finagle.ServiceFactory
 import com.twitter.finagle.Stack
 import com.twitter.finagle.param
 import com.twitter.finagle._
+import com.twitter.finagle.service.Retries.WithdrawOnlyRetryBudget
 import com.twitter.util.tunable.Tunable
 import com.twitter.util.Closable
 import com.twitter.util.CloseOnce
@@ -539,6 +542,14 @@ final class MethodBuilder[Req, Rep] private[finagle] (
   }
 
   private[this] def materialize(): Unit = {
+    val params = if (this.config.retry.responseClassifier eq MethodBuilderRetry.Disabled) {
+      this.params
+    } else {
+      // add a withdrawonlybudget in the stack to prevent double deposit when retries are enabled on the methodbuilder
+      val currentBudget = this.params[Retries.Budget]
+      val wrappedBudget: RetryBudget = new WithdrawOnlyRetryBudget(currentBudget.retryBudget)
+      this.params + Retries.Budget(wrappedBudget, currentBudget.requeueBackoffs)
+    }
     methodPool.materialize(params)
   }